chore: move help logic to the help drawer
